- Restruct the modules, if a module only contains one __init.py__ file like finalize_context/__ini__.py, restructure to finalize_context.py file. - Put all tests under HST/tests directory - In pipeline_query_hst_products & pipeline_query_hst_moving_targets, add a new argument --tq to determine if the scripts should queue the next task.
