+++ HST/finalize_schema.py
