+++ HST/run_coverage.sh
@@ -1,3 +1,3 @@
-coverage run -m pytest */tests
+coverage run -m pytest tests
 coverage html
 coverage report
