+++ HST/tests/test_schema_label.golden.xml
